Refactoring Credentials
