+++ app/src/main/java/ve/com/teeac/svgs/authentication/data/models/Credentials.kt
@@ -1,0 +1,3 @@
+package ve.com.teeac.svgs.authentication.data.models
+
+data class Credentials(val idToken: String, val accessToken: String?)
