+++ app/src/main/java/ve/com/teeac/svgs/authentication/presentation/twitter_button/TwitterButton.kt
@@ -11,7 +11,6 @@ import androidx.compose.ui.platform.LocalContext
 import androidx.compose.ui.res.colorResource
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import androidx.hilt.navigation.compose.hiltViewModel
 import ve.com.teeac.svgs.R
@@ -48,10 +47,4 @@ fun TwitterButton(
         Spacer(modifier = Modifier.width(8.dp))
         Text(stringResource(R.string.auth_with_twitter))
     }
-}
-
-@Preview
-@Composable
-private fun TwitterButtonPreview() {
-    TwitterButton({}, isDisable = false)
 }
