+++ app/src/test/java/ve/com/teeac/svgs/authentication/domain/use_case/SignInWithGoogleUseCaseTest.kt
@@ -13,7 +13,7 @@ import org.junit.Assert.assertEquals
 import org.junit.Before
 import org.junit.Test
 import ve.com.teeac.svgs.authentication.data.data_source.AuthRemoteUser
-import ve.com.teeac.svgs.authentication.data.data_source.Credentials
+import ve.com.teeac.svgs.authentication.data.models.Credentials
 import ve.com.teeac.svgs.authentication.data.models.User
 import ve.com.teeac.svgs.authentication.data.repository.AuthRepositoryImpl
 import ve.com.teeac.svgs.authentication.domain.repositories.AuthRepository
